fix: correct signal type cast and replace Default::default() with explicit initialization in userspace
- Cast SIGINT to i32 in shell sigaction call to match expected syscall signature
- Add shutdown_flags::NORMAL constant mapping to LINUX_REBOOT_CMD_POWER_OFF for compatibility
- Replace Stat::default() with unsafe zeroed() in stat_test to avoid Default trait requirement
- Replace Timespec::default() with explicit struct initialization in time_test for clarity

+++ userspace/shell/src/main.rs
@@ -237,7 +237,7 @@ pub extern "C" fn shell_entry() -> ! {
         libsyscall::sigreturn()
     }
     libsyscall::sigaction(
-        libsyscall::SIGINT,
+        libsyscall::SIGINT as i32,
         sigint_handler as *const () as usize,
         sigreturn_trampoline as *const () as usize,
     );
